+++ src/tempestApi.ts
@@ -1,5 +1,9 @@
 import { Logger } from 'homebridge';
 import axios, { AxiosResponse } from 'axios';
+import https from 'https';
+
+axios.defaults.timeout = 10000; // same as default interval
+axios.defaults.httpsAgent = new https.Agent({ keepAlive: true });
 
 export interface Observation {
   // temperature sensors
@@ -16,7 +20,7 @@ export interface Observation {
   wind_gust: number;               // m/s, used for motion sensor
 
   // occupancy sensors
-  barometric_pressure: number;     // mb
+  barometric_pressure: number;     // mbar
   precip: number;                  // mm/min (minute sampling)
   precip_accum_local_day: number;  // mm
   wind_direction: number;          // degrees
@@ -49,44 +53,34 @@ export class TempestApi {
 
   }
 
-  private async getStationObservation() {
+  private async getStationObservation(): Promise<AxiosResponse | undefined> {
 
-    try {
-      const url = `https://swd.weatherflow.com/swd/rest/observations/station/${this.station_id}`;
-      const options = {
-        headers: {
-          'Authorization': `Bearer ${this.token}`,
-        },
-        validateStatus: (status: number) => status < 500, // Resolve only if the status code is less than 500
-      };
-      return await axios.get(url, options);
-    } catch(exception) {
-      this.log.debug(`[WeatherFlow] ${exception}`);
-      return;
-    }
+    let observation: AxiosResponse | undefined;
+
+    const url = `https://swd.weatherflow.com/swd/rest/observations/station/${this.station_id}`;
+    const options = {
+      headers: {
+        'Authorization': `Bearer ${this.token}`,
+      },
+      validateStatus: (status: number) => status >= 200 && status < 300, // Default
+    };
+
+    await axios.get(url, options)
+      .then(response => {
+        observation = response.data['obs'][0];
+      })
+
+      .catch(exception => {
+        this.log.warn(`[WeatherFlow] ${exception}`);
+      });
+
+    return observation;
 
   }
 
   private async delay(ms: number): Promise<unknown> {
 
     return new Promise(resolve => setTimeout(resolve, ms));
-
-  }
-
-  private isResponseGood(response: AxiosResponse): boolean {
-
-    try {
-      if (!response || !response.data) {
-        return false;
-      } else if (typeof response.data === 'string') {
-        return ('obs' in JSON.parse(response.data));
-      } else {
-        return ('obs' in response.data);
-      }
-    } catch(exception) {
-      this.log.error(exception as string);
-      return false;
-    }
 
   }
 
@@ -97,23 +91,27 @@ export class TempestApi {
       return;
     }
 
-    const response = await this.getStationObservation();
-    if (!response || !this.isResponseGood(response)) {
+    const observation = await this.getStationObservation();
+
+    if (observation === undefined) {
       this.log.warn('Response missing "obs" data.');
+
       if (this.data !== undefined) {
         this.log.warn('Returning last cached response.');
         return this.data;
+
+      } else {
+        this.log.warn(`Retrying ${retry_count + 1} of ${this.max_retries}. No cached "obs" data.`);
+        retry_count += 1;
+        await this.delay(1000 * retry_count);
+        return this.getStationCurrentObservation(retry_count);
       }
-      this.log.warn(`Retrying ${retry_count + 1} of ${this.max_retries}. No cached "obs" data.`);
-      retry_count += 1;
-      await this.delay(1000 * retry_count);
-      return await this.getStationCurrentObservation(retry_count);
+
     } else {
-      if (typeof response.data === 'string') {
-        response.data = JSON.parse(response.data);
-      }
-      this.data = response.data['obs'][0];
+
+      this.data = observation;
       return this.data;
+
     }
 
   }
@@ -125,7 +123,7 @@ export class TempestApi {
       headers: {
         'Authorization': `Bearer ${this.token}`,
       },
-      validateStatus: (status: number) => status < 500, // Resolve only if the status code is less than 500
+      validateStatus: (status: number) => status >= 200 && status < 300, // Default
     };
 
     await axios.get(url, options) // assumes single Tempest station
@@ -135,7 +133,7 @@ export class TempestApi {
       })
 
       .catch(exception => {
-        this.log.debug(`[WeatherFlow] ${exception}`);
+        this.log.warn(`[WeatherFlow] ${exception}`);
       });
 
     return this.tempest_battery_level;
@@ -149,7 +147,7 @@ export class TempestApi {
       headers: {
         'Authorization': `Bearer ${this.token}`,
       },
-      validateStatus: (status: number) => status < 500, // Resolve only if the status code is less than 500
+      validateStatus: (status: number) => status >= 200 && status < 300, // Default
     };
 
     await axios.get(url, options) // assumes single hub with single Tempest station
@@ -158,8 +156,9 @@ export class TempestApi {
       })
 
       .catch(exception => {
-        this.log.debug(`[WeatherFlow] ${exception}`);
+        this.log.warn(`[WeatherFlow] ${exception}`);
       });
+
     return this.tempest_device_id;
 
   }